Move TestLoggers methods to Loggers
This commit migrates the methods TestLoggers#addAppender and
TestLoggers#removeAppenders to the class Loggers as these methods will
be needed in production code.
