--- test/framework/src/main/java/org/elasticsearch/common/logging/TestLoggers.java
@@ -1,57 +1,0 @@
-/*
- * Licensed to Elasticsearch under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.elasticsearch.common.logging;
-
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.apache.logging.log4j.core.Appender;
-import org.apache.logging.log4j.core.LoggerContext;
-import org.apache.logging.log4j.core.config.AppenderRef;
-import org.apache.logging.log4j.core.config.Configuration;
-import org.apache.logging.log4j.core.config.LoggerConfig;
-
-public class TestLoggers {
-
-    public static void addAppender(final Logger logger, final Appender appender) {
-        final LoggerContext ctx = (LoggerContext) LogManager.getContext(false);
-        final Configuration config = ctx.getConfiguration();
-        config.addAppender(appender);
-        LoggerConfig loggerConfig = config.getLoggerConfig(logger.getName());
-        if (!logger.getName().equals(loggerConfig.getName())) {
-            loggerConfig = new LoggerConfig(logger.getName(), logger.getLevel(), true);
-            config.addLogger(logger.getName(), loggerConfig);
-        }
-        loggerConfig.addAppender(appender, null, null);
-        ctx.updateLoggers();
-    }
-
-    public static void removeAppender(final Logger logger, final Appender appender) {
-        final LoggerContext ctx = (LoggerContext) LogManager.getContext(false);
-        final Configuration config = ctx.getConfiguration();
-        LoggerConfig loggerConfig = config.getLoggerConfig(logger.getName());
-        if (!logger.getName().equals(loggerConfig.getName())) {
-            loggerConfig = new LoggerConfig(logger.getName(), logger.getLevel(), true);
-            config.addLogger(logger.getName(), loggerConfig);
-        }
-        loggerConfig.removeAppender(appender.getName());
-        ctx.updateLoggers();
-    }
-
-}
